boxStyles refactor
