+++ design-system/packages/box/src/Box.tsx
@@ -1,108 +1,11 @@
 /** @jsx jsx */
 import { jsx } from '@emotion/core';
-import { useMediaQuery, forwardRefWithAs } from '@nathsimpson/utils';
-import { ReactNode, ElementType } from 'react';
-
-import {
-  BoxMarginProps,
-  BoxPaddingProps,
-  BoxRadiiProps,
-  getMarginStyles,
-  getPaddingStyles,
-  getRadiiStyles,
-  useBoxColors
-} from './utils';
+import { forwardRefWithAs } from '@nathsimpson/utils';
+import { BoxProps, boxStyles } from './styles';
 
 export const Box = forwardRefWithAs<'div', BoxProps>(
-  (
-    {
-      as: Tag = 'div',
-      bg,
-      height,
-      margin,
-      marginBottom,
-      marginLeft,
-      marginRight,
-      marginTop,
-      marginX,
-      marginY,
-      padding,
-      paddingBottom,
-      paddingLeft,
-      paddingRight,
-      paddingTop,
-      paddingX,
-      paddingY,
-      rounding,
-      roundingBottom,
-      roundingLeft,
-      roundingRight,
-      roundingTop,
-      width,
-      ...props
-    },
-    ref
-  ) => {
-    const boxStyles = useBoxStyles({
-      bg,
-      height,
-      margin,
-      marginBottom,
-      marginLeft,
-      marginRight,
-      marginTop,
-      marginX,
-      marginY,
-      padding,
-      paddingBottom,
-      paddingLeft,
-      paddingRight,
-      paddingTop,
-      paddingX,
-      paddingY,
-      rounding,
-      roundingBottom,
-      roundingLeft,
-      roundingRight,
-      roundingTop,
-      width
-    });
-    return <Tag css={boxStyles} ref={ref} {...props} />;
+  ({ as: Tag = 'div', ...props }, ref) => {
+    const [styles, attrs] = boxStyles(props);
+    return <Tag ref={ref} css={styles} {...attrs} />;
   }
 );
-
-export const useBoxStyles = ({
-  bg = 'none',
-  height,
-  width,
-  ...props
-}: BoxProps) => {
-  const { mq, mapResponsiveProp } = useMediaQuery();
-  const margin = getMarginStyles(props);
-  const padding = getPaddingStyles(props);
-  const radii = getRadiiStyles(props);
-  const colors = useBoxColors();
-
-  return mq({
-    background: mapResponsiveProp(bg, colors),
-    height,
-    ...margin,
-    ...padding,
-    ...radii,
-    width
-  });
-};
-
-export type BoxProps = BoxMarginProps &
-  BoxPaddingProps &
-  BoxRadiiProps & {
-    as?: ElementType;
-    /** Background color */
-    bg?: 'none' | 'base' | 'emphasis';
-    /** The content of this container. */
-    children?: ReactNode | ReactNode[] | string;
-    /** Height */
-    height?: number;
-    /** Width */
-    width?: number;
-  };
