+++ design-system/packages/box/src/index.tsx
@@ -1,3 +1,3 @@
 export * from './Box';
 export { Flex } from './Flex';
-export * from './utils';
+export * from './styles';
